# Conflicts:
#	DesktopEditor/graphics/pro/js/wasm/src/drawingfile.cpp

+++ MsBinaryFile/DocFile/CommentsMapping.cpp
@@ -121,7 +121,7 @@ namespace DocFileFormat
 					{
 						if ((m_document->AnnotationsReferencesEx) && (index < m_document->AnnotationsReferencesEx->m_ReferencesEx.size()))
 						{
-							_paraId = XmlUtils::ToString(x++, L"%08X");
+							_paraId = XmlUtils::ToString((unsigned int)(x++), L"%08X");
 						}
 						cp = writeParagraph(cp, 0x7fffffff);
 					}
